requirements.txt
flask~=2.2.2
flask_login
flask_sqlalchemy
flask_mqtt
sqlalchemy~=1.4.46
Werkzeug~=2.2.2
pandas~=1.5.2
paho-mqtt~=1.6.1